Moving simepl_UV_dataset to generated.py
As suggested by @VeckoTheGecko
